scripts/requirements-automation.txt
# Minimal requirements for task automation scripts
PyGithub==2.1.1
python-dateutil==2.8.2 